[all] Update to `avm1-types@0.10`
This commit updates to the latest `avm1-types` version.

The main change is the new `CfgBlock` type. It is now a struct, and the various types are represent with the field `flow` with the type `CfgFlow`.
The structure of the types was also revamped so `raw` and `cfg` actions are equally supported.